+++ docs/CHANGELOG.md
@@ -31,6 +31,7 @@ Version 1.8.1 (not released yet)
 - Shorten the wait between levels when the server is empty
 - Support CIDR IP ranges in banlist (should help with banning VPNs)
 - Fix alpha sorting regression in level L5S3 (and possibly others)
+- Add more info to `level_info` command and add `map_info` alias
 
 Version 1.8.0 (released 2022-09-17)
 -----------------------------------